# _refs/openapi-soundcloud-com.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "SoundCloud Public API Specification",
    "description": "Swagger json can be found [here](https://developers.soundcloud.com/docs/api/explorer/api.json)\n\n\u003Ca href=\"#\" id=\"cookie-manager\"\u003ECookie manager\u003C/a\u003E\n",
    "contact": {
      "name": "API issue tracker",
      "url": "https://github.com/soundcloud/api"
    },
    "version": "1.0.0",
    "termsOfService": "https://developers.soundcloud.com/docs/api/terms-of-use"
  },
  "servers": [
    {
      "url": "https://api.soundcloud.com"
    }
  ],
  "security": [
    {
      "AuthHeader": []
    },
    {
      "oAuth2_1": []
    }
  ],
...
  "paths": {
...
    "/playlists/{playlist_urn}": {
      "get": {
        "summary": "Returns a playlist.",
        "tags": [
          "playlists"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/playlist_urn"
          },
          {
            "$ref": "#/components/parameters/secret_token"
          },
          {
            "$ref": "#/components/parameters/access"
          },
          {
            "$ref": "#/components/parameters/show_tracks"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Playlist"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "put": {
        "summary": "Updates a playlist.",
        "tags": [
          "playlists"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/fake_playlist_urn"
          }
        ],
        "requestBody": {
          "description": "Playlist payload",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUpdatePlaylistRequest"
              },
              "examples": {
                "default": {
                  "$ref": "#/components/examples/CreateUpdatePlaylistRequest"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Playlist"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "summary": "Deletes a playlist.",
        "tags": [
          "playlists"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/fake_playlist_urn"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/tracks/{track_urn}": {
      "get": {
        "summary": "Returns a track.",
        "tags": [
          "tracks"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/track_urn"
          },
          {
            "$ref": "#/components/parameters/secret_token"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Track"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "summary": "Updates a track's information.",
        "tags": [
          "tracks"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/track_urn"
          }
        ],
        "requestBody": {
          "description": "Track payload",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackMetadataRequest"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/TrackDataRequest"
              }
            },
            "multipart/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/TrackDataRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Track"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "delete": {
        "summary": "Deletes a track.",
        "tags": [
          "tracks"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/track_urn"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/likes/tracks/{track_urn}": {
      "post": {
        "summary": "Likes a track.",
        "tags": [
          "likes"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/test_track_urn"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      },
      "delete": {
        "summary": "Unlikes a track.",
        "tags": [
          "likes"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/test_track_urn"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/likes/playlists/{playlist_urn}": {
      "post": {
        "summary": "Likes a playlist.",
        "tags": [
          "likes"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/playlist_urn"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      },
      "delete": {
        "summary": "Unlikes a playlist.",
        "tags": [
          "likes"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/playlist_urn"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "secret_token": {
        "in": "query",
        "name": "secret_token",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "A secret token to fetch private playlists/tracks"
      },
      "access": {
        "in": "query",
        "name": "access",
        "schema": {
          "type": "array",
          "items": {
            "type": "string",
            "enum": [
              "playable",
              "preview",
              "blocked"
            ]
          },
          "default": "playable,preview"
        },
        "description": "Filters content by level of access the user (logged in or anonymous) has to the track. The result list will include only tracks with the specified access. Include all options if you'd like to see all possible tracks. See `Track#access` schema for more details.\n",
        "style": "form",
        "explode": false
      },
...
      "show_tracks": {
        "in": "query",
        "name": "show_tracks",
        "required": false,
        "schema": {
          "type": "boolean"
        },
        "description": "A boolean flag to request a playlist with or without tracks. Default is `true`.",
        "example": true
      },
...
      "playlist_urn": {
        "in": "path",
        "name": "playlist_urn",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "SoundCloud playlist id",
        "example": "soundcloud:playlists:1212781357"
      },
      "fake_playlist_urn": {
        "in": "path",
        "name": "playlist_urn",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "SoundCloud playlist id",
        "example": "soundcloud:playlists:10"
      },
...
      "track_urn": {
        "in": "path",
        "name": "track_urn",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "SoundCloud Track urn",
        "example": "soundcloud:tracks:308946187"
      },
      "test_track_urn": {
        "in": "path",
        "name": "track_urn",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "SoundCloud Track id",
        "example": "soundcloud:tracks:1015448728"
      },
...
    },
    "securitySchemes": {
      "AuthHeader": {
        "in": "header",
        "type": "apiKey",
        "name": "Authorization",
        "description": "The string \"OAuth\" followed by the user’s token: OAuth ACCESS_TOKEN. Obtained in oAuth2.1 flow"
      },
      "oAuth2_1": {
        "type": "oauth2",
        "description": "This API uses OAuth2.1 with Authorization Code flow, Client Credential flow. When registering your app, for it to work with Swagger UI specify redirect url as https://developers.soundcloud.com/docs/api/explorer/oauth2-redirect.html. [More info](https://developers.soundcloud.com/docs/api/guide#authentication)",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://secure.soundcloud.com/authorize",
            "tokenUrl": "https://secure.soundcloud.com/oauth/token",
            "refreshUrl": "https://secure.soundcloud.com/oauth/token",
            "scopes": {

            }
          },
          "clientCredentials": {
            "tokenUrl": "https://secure.soundcloud.com/oauth/token",
            "scopes": {

            }
          }
        }
      }
    },
    "schemas": {
...
      "TrackMetadataRequest": {
        "type": "object",
        "properties": {
          "track": {
            "type": "object",
            "properties": {
              "title": {
                "type": "string"
              },
              "permalink": {
                "type": "string"
              },
              "sharing": {
                "type": "string",
                "enum": [
                  "public",
                  "private"
                ],
                "default": "public"
              },
              "embeddable_by": {
                "type": "string",
                "enum": [
                  "all",
                  "me",
                  "none"
                ],
                "description": "who can embed this track \"all\", \"me\", or \"none\""
              },
              "purchase_url": {
                "type": "string"
              },
              "description": {
                "type": "string"
              },
              "genre": {
                "type": "string"
              },
              "tag_list": {
                "type": "string",
                "description": "The tag_list property contains a list of tags separated by spaces. Multiword tags are quoted in double quotes. We also support machine tags that follow the pattern NAMESPACE:KEY=VALUE. For example: geo:lat=43.555 camel:size=medium “machine:tag=with space” Machine tags are not revealed to the user on the track pages."
              },
              "label_name": {
                "type": "string"
              },
              "release": {
                "type": "string"
              },
              "release_date": {
                "title": "write only!",
                "type": "string",
                "description": "string, formatted as yyyy-mm-dd, representing release date"
              },
              "streamable": {
                "type": "boolean",
                "default": true
              },
              "downloadable": {
                "type": "boolean",
                "default": true
              },
              "license": {
                "type": "string",
                "enum": [
                  "no-rights-reserved",
                  "all-rights-reserved",
                  "cc-by",
                  "cc-by-nc",
                  "cc-by-nd",
                  "cc-by-sa",
                  "cc-by-nc-nd",
                  "cc-by-nc-sa"
                ],
                "description": "Possible values: no-rights-reserved, all-rights-reserved, cc-by, cc-by-nc, cc-by-nd, cc-by-sa, cc-by-nc-nd, cc-by-nc-sa"
              },
              "commentable": {
                "type": "boolean",
                "default": true
              },
              "isrc": {
                "type": "string"
              }
            }
          }
        }
      },
      "CreateUpdatePlaylistRequest": {
        "type": "object",
        "properties": {
          "playlist": {
            "type": "object",
            "properties": {
              "title": {
                "description": "Title of the playlist",
                "type": "string"
              },
              "description": {
                "description": "Description of the playlist",
                "type": "string"
              },
              "sharing": {
                "description": "public or private",
                "type": "string",
                "enum": [
                  "public",
                  "private"
                ]
              },
              "tracks": {
                "description": "List of tracks to add to playlist",
                "type": "array",
                "items": {
                  "type": "object",
                  "required": [
                    "urn"
                  ],
                  "properties": {
                    "urn": {
                      "description": "SoundCloud track urn",
                      "type": "string"
                    }
                  }
                }
              },
              "artwork_data": {
                "title": "Only for POST",
                "type": "string",
                "format": "binary"
              },
              "ean": {
                "description": "The European Article Number",
                "type": "string"
              },
              "genre": {
                "description": "Playlist's genre",
                "type": "string"
              },
              "label_name": {
                "description": "Label name",
                "type": "string"
              },
              "license": {
                "description": "License number",
                "type": "string"
              },
              "permalink": {
                "description": "Playlist's permalink",
                "type": "string"
              },
              "permalink_url": {
                "description": "Full permalink URL",
                "type": "string"
              },
              "purchase_title": {
                "description": "Purchase title",
                "type": "string"
              },
              "purchase_url": {
                "description": "Purchase URL",
                "type": "string"
              },
              "release": {
                "description": "Playlist's release",
                "type": "string"
              },
              "release_date": {
                "description": "Release date",
                "type": "string"
              },
              "set_type": {
                "description": "Playlist or album type",
                "type": "string",
                "enum": [
                  "album",
                  "playlist"
                ]
              },
              "tag_list": {
                "description": "A comma-separated list of tags",
                "type": "string"
              }
            }
          }
        }
      },
...
      "TrackDataRequest": {
        "type": "object",
        "properties": {
          "track[title]": {
            "type": "string"
          },
          "track[asset_data]": {
            "title": "Only for uploading",
            "type": "string",
            "format": "binary"
          },
          "track[permalink]": {
            "type": "string"
          },
          "track[sharing]": {
            "type": "string",
            "enum": [
              "public",
              "private"
            ],
            "default": "public"
          },
          "track[embeddable_by]": {
            "type": "string",
            "enum": [
              "all",
              "me",
              "none"
            ],
            "description": "who can embed this track \"all\", \"me\", or \"none\""
          },
          "track[purchase_url]": {
            "type": "string"
          },
          "track[description]": {
            "type": "string"
          },
          "track[genre]": {
            "type": "string"
          },
          "track[tag_list]": {
            "type": "string",
            "description": "The tag_list property contains a list of tags separated by spaces. Multiword tags are quoted in double quotes. We also support machine tags that follow the pattern NAMESPACE:KEY=VALUE. For example: geo:lat=43.555 camel:size=medium “machine:tag=with space” Machine tags are not revealed to the user on the track pages."
          },
          "track[label_name]": {
            "type": "string"
          },
          "track[release]": {
            "type": "string"
          },
          "track[release_date]": {
            "title": "write only!",
            "type": "string",
            "description": "string, formatted as yyyy-mm-dd, representing release date"
          },
          "track[streamable]": {
            "type": "boolean",
            "default": true
          },
          "track[downloadable]": {
            "type": "boolean",
            "default": true
          },
          "track[license]": {
            "type": "string",
            "enum": [
              "no-rights-reserved",
              "all-rights-reserved",
              "cc-by",
              "cc-by-nc",
              "cc-by-nd",
              "cc-by-sa",
              "cc-by-nc-nd",
              "cc-by-nc-sa"
            ],
            "description": "Possible values: no-rights-reserved, all-rights-reserved, cc-by, cc-by-nc, cc-by-nd, cc-by-sa, cc-by-nc-nd, cc-by-nc-sa"
          },
          "track[commentable]": {
            "type": "boolean",
            "default": true
          },
          "track[isrc]": {
            "type": "string"
          },
          "track[artwork_data]": {
            "title": "Only for uploading, for PRO users",
            "type": "string",
            "format": "binary"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "link": {
            "type": "string"
          },
          "error": {
            "type": "string",
            "deprecated": true,
            "nullable": true
          },
          "errors": {
            "type": "array",
            "deprecated": true,
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "deprecated": true
          }
        }
      },
      "TooManyRequests": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Error"
          },
          {
            "type": "object",
            "properties": {
              "spam_warning_urn": {
                "type": "string"
              }
            }
          }
        ]
      },
...
      "Playlist": {
        "type": "object",
        "description": "Soundcloud Playlist Object",
        "properties": {
          "title": {
            "description": "Playlist title.",
            "type": "string"
          },
          "urn": {
            "description": "Playlist identifier.",
            "type": "string"
          },
          "kind": {
            "description": "Type of Soundcloud object (playlist).",
            "type": "string"
          },
          "artwork_url": {
            "description": "URL to a JPEG image.",
            "type": "string"
          },
          "created_at": {
            "description": "Created timestamp.",
            "type": "string"
          },
          "description": {
            "description": "Playlist description.",
            "type": "string"
          },
          "downloadable": {
            "description": "is downloadable.",
            "type": "boolean"
          },
          "duration": {
            "description": "Playlist duration.",
            "type": "integer"
          },
          "ean": {
            "description": "European Article Number.",
            "type": "string"
          },
          "embeddable_by": {
            "description": "Embeddable by.",
            "type": "string"
          },
          "genre": {
            "description": "Playlist genre.",
            "type": "string"
          },
          "label_id": {
            "description": "Label user identifier.",
            "type": "integer"
          },
          "label_name": {
            "description": "Label name.",
            "type": "string"
          },
          "last_modified": {
            "description": "Last modified timestamp.",
            "type": "string"
          },
          "license": {
            "description": "License.",
            "type": "string"
          },
          "permalink": {
            "description": "Playlist permalink.",
            "type": "string"
          },
          "permalink_url": {
            "description": "Playlist permalink URL.",
            "type": "string"
          },
          "playlist_type": {
            "description": "Type of playlist.",
            "type": "string"
          },
          "purchase_title": {
            "description": "Purchase title.",
            "type": "string"
          },
          "purchase_url": {
            "description": "Purchase URL.",
            "type": "string"
          },
          "release": {
            "description": "Release.",
            "type": "string"
          },
          "release_day": {
            "description": "Day of release.",
            "type": "integer"
          },
          "release_month": {
            "description": "Month of release.",
            "type": "integer"
          },
          "release_year": {
            "description": "Year of release.",
            "type": "integer"
          },
          "sharing": {
            "description": "Type of sharing (private/public).",
            "type": "string"
          },
          "streamable": {
            "description": "Is streamable.",
            "type": "boolean"
          },
          "tag_list": {
            "description": "Tags.",
            "type": "string"
          },
          "track_count": {
            "description": "Count of tracks.",
            "type": "integer"
          },
          "tracks": {
            "description": "List of tracks.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Track"
            }
          },
          "type": {
            "description": "Playlist type.",
            "type": "string"
          },
          "uri": {
            "description": "Playlist URI.",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "user_urn": {
            "description": "User identifier.",
            "type": "string"
          },
          "likes_count": {
            "description": "Count of playlist likes.",
            "type": "integer"
          },
          "label": {
            "allOf": [
              {
                "type": "object",
                "nullable": true
              },
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "tracks_uri": {
            "description": "tracks URI.",
            "type": "string",
            "nullable": true
          },
          "tags": {
            "description": "Tags.",
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "responses": {
      "Success": {
        "description": "Success",
        "content": {
          "application/json; charset=utf-8": {
            "examples": {
              "default": {
                "$ref": "#/components/examples/Success"
              }
            }
          }
        }
      },
...
      "BadRequest": {
        "description": "Bad Request",
        "content": {
          "application/json; charset=utf-8": {
            "examples": {
              "default": {
                "$ref": "#/components/examples/BadRequest"
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json; charset=utf-8": {
            "examples": {
              "default": {
                "$ref": "#/components/examples/Unauthorized"
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Not Found",
        "content": {
          "application/json": {
            "examples": {
              "default": {
                "$ref": "#/components/examples/NotFound"
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "TooManyRequests": {
        "description": "Too Many Requests",
        "content": {
          "application/json": {
            "examples": {
              "default": {
                "$ref": "#/components/examples/TooManyRequests"
              }
            },
            "schema": {
              "$ref": "#/components/schemas/TooManyRequests"
            }
          }
        }
      },
...
      "Track": {
        "description": "Success",
        "content": {
          "application/json; charset=utf-8": {
            "schema": {
              "$ref": "#/components/schemas/Track"
            },
            "examples": {
              "default": {
                "$ref": "#/components/examples/Track"
              }
            }
          }
        }
      },
...
      "Playlist": {
        "description": "Success",
        "content": {
          "application/json; charset=utf-8": {
            "schema": {
              "$ref": "#/components/schemas/Playlist"
            },
            "examples": {
              "default": {
                "$ref": "#/components/examples/Playlist"
              }
            }
          }
        }
      },
...
    },
    "examples": {
...
      "CreateUpdatePlaylistRequest": {
        "value": {
          "playlist": {
            "title": "Test Auto-created Playlist",
            "description": "Playlist for test purposes",
            "sharing": "private",
            "tracks": [
              {
                "urn": "soundcloud:tracks:219787221"
              },
              {
                "urn": "soundcloud:tracks:783019264"
              },
              {
                "urn": "soundcloud:tracks:870073492"
              }
            ]
          }
        }
      },
...
      "Playlist": {
        "value": {
          "artwork_url": "",
          "created_at": "2020/07/16 16:09:54 +0000",
          "description": null,
          "downloadable": true,
          "duration": 11000,
          "ean": null,
          "embeddable_by": "all",
          "genre": "",
          "urn": "soundcloud:playlists:12345",
          "kind": "playlist",
          "label": null,
          "label_id": null,
          "label_name": null,
          "last_modified": "2020/09/08 09:13:00 +0000",
          "license": "all-rights-reserved",
          "likes_count": 0,
          "permalink": "permalink",
          "permalink_url": "https://soundcloud.com/userPermalink/sets/permalink",
          "playlist_type": "",
          "purchase_title": null,
          "purchase_url": null,
          "release": null,
          "release_day": null,
          "release_month": null,
          "release_year": null,
          "sharing": "public",
          "streamable": true,
          "tag_list": "",
          "tags": "",
          "title": "permalink",
          "track_count": 2,
          "tracks": {
            "allOf": [
              {
                "$ref": "#/components/examples/Track/value"
              }
            ]
          },
          "tracks_uri": "https://api.soundcloud.com/playlists/soundcloud:playlists:12345/tracks",
          "type": "",
          "uri": "https://api.soundcloud.com/playlists/soundcloud:playlists:12345",
          "user": {
            "$ref": "#/components/examples/User/value"
          },
          "user_id": 1234
        }
      },
...
      "Track": {
        "value": {
          "artwork_url": "https://i1.sndcdn.com/artworks-large.jpg",
          "available_country_codes": null,
          "bpm": null,
          "comment_count": 4,
          "commentable": true,
          "created_at": "2019/01/15 15:40:49 +0000",
          "description": null,
          "download_count": 0,
          "download_url": "https://api.soundcloud.com/tracks/1234/download",
          "downloadable": false,
          "duration": 40000,
          "embeddable_by": "all",
          "favoritings_count": 271,
          "genre": "Rock",
          "urn": "soundcloud:tracks:1234",
          "isrc": null,
          "key_signature": null,
          "kind": "track",
          "label_name": "some label",
          "license": "all-rights-reserved",
          "permalink_url": "https://soundcloud.com/userPermalink/trackPermalink",
          "playback_count": 8027,
          "purchase_title": null,
          "purchase_url": null,
          "release": null,
          "release_day": 22,
          "release_month": 8,
          "release_year": 2019,
          "reposts_count": 18,
          "secret_uri": null,
          "sharing": "public",
          "stream_url": "https://api.soundcloud.com/tracks/1234/stream",
          "streamable": true,
          "tag_list": "",
          "title": "Some title",
          "uri": "https://api.soundcloud.com/tracks/1234",
          "user": {
            "$ref": "#/components/examples/User/value"
          },
          "user_favorite": true,
          "user_playback_count": 1,
          "waveform_url": "https://wave.sndcdn.com/someString.png",
          "access": "playable"
        }
      },
...
      "Success": {
        "value": {
          "status": "200 - OK"
        }
      },
...
      "BadRequest": {
        "value": {
          "code": 400,
          "message": "Detailed message of errors, when available",
          "link": "https://developers.soundcloud.com/docs/api/explorer/open-api"
        }
      },
...
      "Unauthorized": {
        "value": {
          "code": 401,
          "message": "Detailed message of errors, when available",
          "link": "https://developers.soundcloud.com/docs/api/explorer/open-api"
        }
      },
...
      "NotFound": {
        "value": {
          "code": 404,
          "message": "Detailed message of errors, when available",
          "link": "https://developers.soundcloud.com/docs/api/explorer/open-api"
        }
      },
...
      "TooManyRequests": {
        "value": {
          "code": 429,
          "message": "Detailed message of errors, when available",
          "spam_warning_urn": "soundcloud:spam-warnings:1",
          "link": "https://developers.soundcloud.com/docs/api/rate-limits#errors"
        }
      },
...
    }
  }
}